don't delete chats from recent search

+++ TMessagesProj/src/main/java/org/telegram/messenger/MessagesStorage.java
@@ -10754,9 +10754,31 @@ public class MessagesStorage extends BaseController {
 
     public void clearFileProtectedDb(SQLiteDatabase db) {
         storageQueue.postRunnable(() -> {
+            SQLiteCursor cursor = null;
             try {
+                cursor = db.queryFinalized("SELECT did FROM search_recent WHERE 1");
+                Set<Long> recentSearchDialogIds = new HashSet<>();
+                while (cursor.next()) {
+                    recentSearchDialogIds.add(cursor.longValue(0));
+                }
+                cursor.dispose();
+                cursor = null;
+                cursor = db.queryFinalized("SELECT uid FROM chats WHERE 1");
+                Set<Long> chatIdsToDelete = new HashSet<>();
+                while (cursor.next()) {
+                    long chatId = cursor.longValue(0);
+                    if (!recentSearchDialogIds.contains(chatId)) {
+                        chatIdsToDelete.add(chatId);
+                    }
+                }
+                cursor.dispose();
+                cursor = null;
+
+                for (Long chatId : recentSearchDialogIds) {
+                    db.executeFast("DELETE FROM chats WHERE uid = " + chatId).stepThis().dispose();
+                }
+
                 String notEncryptedGroupCheck = "did & 0x4000000000000000 = 0 OR did & 0x8000000000000000 <> 0";
-                db.executeFast("DELETE FROM chats").stepThis().dispose();
                 db.executeFast("DELETE FROM contacts").stepThis().dispose();
                 db.executeFast("DELETE FROM messages_v2 WHERE " + notEncryptedGroupCheck.replace("did", "uid")).stepThis().dispose();
                 db.executeFast("DELETE FROM dialogs WHERE " + notEncryptedGroupCheck).stepThis().dispose();
@@ -10766,6 +10788,9 @@ public class MessagesStorage extends BaseController {
             } catch (Exception e) {
                 checkSQLException(e);
             } finally {
+                if (cursor != null) {
+                    cursor.dispose();
+                }
                 AndroidUtilities.runOnUIThread(() -> getNotificationCenter().postNotificationName(NotificationCenter.onFileProtectedDbCleared));
             }
         });
